Move tunnel-only killswitch logic into protocol modules
Refactor the fail-closed killswitch for protocols whose clients do not
implement one (Warp, OpenConnect, OpenFortiVpn) out of the exec CLI into
the respective vopono_core modules. Each protocol now exposes an
apply_killswitch function carrying its own server resolution, tunnel
interface prefix and control endpoint, and the shared server resolution
lives in firewall.rs.
